chore(package): Release 0.2.0 This release adds ability to specify Vimeo iFrame URL parameters. Thanks to @redroot
While only a feature addition and no breaking changes to the API
have been added, we've updated from babel 5 to babel 6, which
brings with it some internal changes to how the library files
are transpiled. While this shouldn't break anything, a minor bump
(which is a major bump when major is 0) is used instead.

Happy Coding.